backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.1.0
gunicorn==22.0.0
python-dotenv==1.0.1
